backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
marshmallow==3.20.1
python-telegram-bot==20.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apscheduler==3.10.4
gunicorn==21.2.0
eventlet